+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.29.0
+------------------------------
+*February 16, 2018*
+
+### Changed
+- Renamed base `scss` file to `fozzie.scss`
+
 v0.28.0
 ------------------------------
 *February 14, 2018*